refactor: reuse file interaction logic in hooks
