+++ src/components/organisms/FileTreePane/FileTreePane.tsx
@@ -2,48 +2,32 @@ import {ipcRenderer} from 'electron';
 
 import React, {Key, useCallback, useEffect, useMemo, useRef, useState} from 'react';
 import {useSelector} from 'react-redux';
-import {useUpdateEffect} from 'react-use';
 
-import {Button, Modal, Tooltip} from 'antd';
+import {Button, Tooltip} from 'antd';
 
 import {ExclamationCircleOutlined, ReloadOutlined} from '@ant-design/icons';
 
 import log from 'loglevel';
-import micromatch from 'micromatch';
 import path from 'path';
 
 import {DEFAULT_PANE_TITLE_HEIGHT, ROOT_FILE_ENTRY, TOOLTIP_DELAY} from '@constants/constants';
 import {CollapseTreeTooltip, ExpandTreeTooltip, FileExplorerChanged, ReloadFolderTooltip} from '@constants/tooltips';
 
-import {AlertEnum} from '@models/alert';
-
 import {useAppDispatch, useAppSelector} from '@redux/hooks';
-import {setAlert} from '@redux/reducers/alert';
-import {updateProjectConfig} from '@redux/reducers/appConfig';
-import {selectFile, setSelectingFile, updateResourceFilter} from '@redux/reducers/main';
-import {
-  openCreateFileFolderModal,
-  openNewResourceWizard,
-  openRenameEntityModal,
-  setExpandedFolders,
-} from '@redux/reducers/ui';
-import {fileIncludesSelector, isInPreviewModeSelector, scanExcludesSelector, settingsSelector} from '@redux/selectors';
-import {getResourcesForPath} from '@redux/services/fileEntry';
-import {getHelmValuesFile} from '@redux/services/helm';
-import {isKustomizationResource} from '@redux/services/kustomize';
-import {startPreview, stopPreview} from '@redux/services/preview';
+import {setSelectingFile, updateResourceFilter} from '@redux/reducers/main';
+import {openRenameEntityModal, setExpandedFolders, openCreateFileFolderModal} from '@redux/reducers/ui';
+import {isInPreviewModeSelector, settingsSelector} from '@redux/selectors';
 import {setRootFolder} from '@redux/thunks/setRootFolder';
 
 import {TitleBar} from '@molecules';
 
 import {Icon} from '@atoms';
 
-import {duplicateEntity} from '@utils/files';
-import {uniqueArr} from '@utils/index';
+import {useCreate, useDelete, useFileSelect, useHighlightNode, usePreview, useProcessing, useDuplicate} from '@hooks/fileTreeHooks';
 
 import {createNode} from './CreateNode';
 import TreeItem from './TreeItem';
-import {ProcessingEntity, TreeNode} from './types';
+import {TreeNode} from './types';
 
 import * as S from './styled';
 
@@ -52,21 +36,13 @@ type Props = {
 };
 
 const FileTreePane: React.FC<Props> = ({height}) => {
-  const [highlightNode, setHighlightNode] = useState<TreeNode>();
-  const [processingEntity, setProcessingEntity] = useState<ProcessingEntity>({
-    processingEntityID: undefined,
-    processingType: undefined,
-  });
   const [tree, setTree] = useState<TreeNode | null>(null);
-  const leftMenuSelection = useAppSelector(state => state.ui.leftMenu.selection);
   const isInPreviewMode = useSelector(isInPreviewModeSelector);
 
   const dispatch = useAppDispatch();
-  const configState = useAppSelector(state => state.config);
   const expandedFolders = useAppSelector(state => state.ui.leftMenu.expandedFolders);
   const fileMap = useAppSelector(state => state.main.fileMap);
   const fileOrFolderContainedInFilter = useAppSelector(state => state.main.resourceFilter.fileOrFolderContainedIn);
-  const helmValuesMap = useAppSelector(state => state.main.helmValuesMap);
   const isFolderLoading = useAppSelector(state => state.ui.isFolderLoading);
   const isScanExcludesUpdated = useAppSelector(state => state.config.isScanExcludesUpdated);
   const isSelectingFile = useAppSelector(state => state.main.isSelectingFile);
@@ -75,12 +51,20 @@ const FileTreePane: React.FC<Props> = ({height}) => {
   const resourceMap = useAppSelector(state => state.main.resourceMap);
   const selectedPath = useAppSelector(state => state.main.selectedPath);
   const selectedResourceId = useAppSelector(state => state.main.selectedResourceId);
+  const onFileSelect = useFileSelect();
+  const onPreview = usePreview();
+  const {onDelete, processingEntity, setProcessingEntity} = useDelete();
+  const {onCreateResource} = useCreate();
+  const {onDuplicate} = useDuplicate();
+
+  const onCreateFileFolder = (absolutePath: string, type: 'file' | 'folder') => {
+    dispatch(openCreateFileFolderModal({rootDir: absolutePath, type}));
+  };
 
   const {hideExcludedFilesInFileExplorer, hideUnsupportedFilesInFileExplorer} = useAppSelector(settingsSelector);
-  const fileIncludes = useAppSelector(fileIncludesSelector);
-  const scanExcludes = useAppSelector(scanExcludesSelector);
 
   const treeRef = useRef<any>();
+  const highlightFilePath = useHighlightNode(tree, treeRef, expandedFolders);
 
   const isButtonDisabled = !fileMap[ROOT_FILE_ENTRY];
   const isCollapsed = expandedFolders.length === 0;
@@ -99,6 +83,8 @@ const FileTreePane: React.FC<Props> = ({height}) => {
   const refreshFolder = useCallback(() => {
     setFolder(fileMap[ROOT_FILE_ENTRY].filePath);
   }, [fileMap, setFolder]);
+
+  const {onExcludeFromProcessing, onIncludeToProcessing} = useProcessing(refreshFolder);
 
   useEffect(() => {
     if (isFolderLoading) {
@@ -136,32 +122,6 @@ const FileTreePane: React.FC<Props> = ({height}) => {
    * when a resource is selected
    */
 
-  function highlightFilePath(filePath: string) {
-    const paths = filePath.split(path.sep);
-    const keys: Array<React.Key> = [ROOT_FILE_ENTRY];
-
-    for (let c = 1; c < paths.length; c += 1) {
-      keys.push(paths.slice(0, c + 1).join(path.sep));
-    }
-
-    let node: TreeNode | undefined = tree || undefined;
-    for (let c = 1; c < keys.length && node; c += 1) {
-      node = node.children.find((i: any) => i.key === keys[c]);
-    }
-
-    if (node) {
-      node.highlight = true;
-      treeRef?.current?.scrollTo({key: node.key});
-
-      if (highlightNode) {
-        highlightNode.highlight = false;
-      }
-    }
-
-    setHighlightNode(node);
-    dispatch(setExpandedFolders(uniqueArr([...expandedFolders, ...Array.from(keys)])));
-  }
-
   useEffect(() => {
     if (selectedResourceId && tree) {
       const resource = resourceMap[selectedResourceId];
@@ -174,116 +134,8 @@ const FileTreePane: React.FC<Props> = ({height}) => {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [selectedResourceId, tree]);
 
-  const onDelete = (args: {isDirectory: boolean; name: string; err: NodeJS.ErrnoException | null}): void => {
-    const {isDirectory, name, err} = args;
-
-    if (err) {
-      dispatch(
-        setAlert({
-          title: 'Deleting failed',
-          message: `Something went wrong during deleting a ${isDirectory ? 'directory' : 'file'}`,
-          type: AlertEnum.Error,
-        })
-      );
-    } else {
-      dispatch(
-        setAlert({
-          title: `Successfully deleted a ${isDirectory ? 'directory' : 'file'}`,
-          message: `You have successfully deleted ${name} ${isDirectory ? 'directory' : 'file'}`,
-          type: AlertEnum.Success,
-        })
-      );
-    }
-
-    /**
-     * Deleting is performed immediately.
-     * The Ant Tree component is not updated immediately.
-     * I show the loader long enough to let the Ant Tree component update.
-     */
-    setTimeout(() => {
-      setProcessingEntity({processingEntityID: undefined, processingType: undefined});
-    }, 2000);
-  };
-
-  const onDuplicate = (absolutePathToEntity: string, entityName: string, dirName: string) => {
-    duplicateEntity(absolutePathToEntity, entityName, dirName, args => {
-      const {duplicatedFileName, err} = args;
-
-      if (err) {
-        dispatch(
-          setAlert({
-            title: 'Duplication failed',
-            message: `Something went wrong during duplicating "${absolutePathToEntity}"`,
-            type: AlertEnum.Error,
-          })
-        );
-      } else {
-        dispatch(
-          setAlert({
-            title: `Duplication succeded`,
-            message: `You have successfully created ${duplicatedFileName}`,
-            type: AlertEnum.Success,
-          })
-        );
-      }
-    });
-  };
-
   const onRename = (absolutePathToEntity: string, osPlatform: string) => {
     dispatch(openRenameEntityModal({absolutePathToEntity, osPlatform}));
-  };
-
-  const onSelect = (selectedKeysValue: React.Key[], info: any) => {
-    if (!fileIncludes.some(fileInclude => micromatch.isMatch(path.basename(info.node.key), fileInclude))) {
-      return;
-    }
-    if (scanExcludes.some(scanExclude => micromatch.isMatch(path.basename(info.node.key), scanExclude))) {
-      return;
-    }
-    if (info.node.key) {
-      if (isInPreviewMode) {
-        stopPreview(dispatch);
-      }
-      dispatch(setSelectingFile(true));
-      dispatch(selectFile({filePath: info.node.key}));
-    }
-  };
-
-  const openConfirmModal = () => {
-    Modal.confirm({
-      title: 'You should reload the file explorer to have your changes applied. Do you want to do it now?',
-      icon: <ExclamationCircleOutlined />,
-      cancelText: 'Not now',
-      onOk: () => {
-        refreshFolder();
-      },
-    });
-  };
-
-  const onExcludeFromProcessing = (relativePath: string) => {
-    dispatch(
-      updateProjectConfig({
-        config: {
-          ...configState.projectConfig,
-          scanExcludes: [...scanExcludes, relativePath],
-        },
-        fromConfigFile: false,
-      })
-    );
-    openConfirmModal();
-  };
-
-  const onIncludeToProcessing = (relativePath: string) => {
-    dispatch(
-      updateProjectConfig({
-        config: {
-          ...configState.projectConfig,
-          scanExcludes: scanExcludes.filter(scanExclude => scanExclude !== relativePath),
-        },
-        fromConfigFile: false,
-      })
-    );
-    openConfirmModal();
   };
 
   useEffect(() => {
@@ -348,61 +200,6 @@ const FileTreePane: React.FC<Props> = ({height}) => {
     dispatch(setExpandedFolders(isCollapsed ? allTreeKeys : []));
   };
 
-  const onCreateFileFolder = (absolutePath: string, type: 'file' | 'folder') => {
-    dispatch(openCreateFileFolderModal({rootDir: absolutePath, type}));
-  };
-
-  const onPreview = useCallback(
-    (relativePath: string) => {
-      const resources = getResourcesForPath(relativePath, resourceMap);
-      if (resources && resources.length === 1 && isKustomizationResource(resources[0])) {
-        startPreview(resources[0].id, 'kustomization', dispatch);
-      } else {
-        const fileEntry = fileMap[relativePath];
-        if (fileEntry) {
-          const valuesFile = getHelmValuesFile(fileEntry, helmValuesMap);
-          if (valuesFile) {
-            startPreview(valuesFile.id, 'helm', dispatch);
-          }
-        }
-      }
-    },
-    [dispatch, fileMap, helmValuesMap, resourceMap]
-  );
-
-  const onCreateResource = ({targetFolder, targetFile}: {targetFolder?: string; targetFile?: string}) => {
-    if (targetFolder) {
-      dispatch(openNewResourceWizard({defaultInput: {targetFolder}}));
-    }
-    if (targetFile) {
-      dispatch(openNewResourceWizard({defaultInput: {targetFile}}));
-    }
-  };
-
-  useEffect(() => {
-    // removes any highlight when a file is selected
-    if (selectedPath && highlightNode) {
-      highlightNode.highlight = false;
-    }
-
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [highlightNode]);
-
-  useUpdateEffect(() => {
-    if (leftMenuSelection !== 'file-explorer') {
-      return;
-    }
-
-    if (selectedPath) {
-      treeRef?.current?.scrollTo({key: selectedPath});
-      return;
-    }
-
-    if (highlightNode) {
-      treeRef?.current?.scrollTo({key: highlightNode.key});
-    }
-  }, [tree]);
-
   const onFilterByFileOrFolder = (relativePath: string | undefined) => {
     dispatch(updateResourceFilter({...resourceFilter, fileOrFolderContainedIn: relativePath}));
   };
@@ -451,7 +248,7 @@ const FileTreePane: React.FC<Props> = ({height}) => {
           </S.RootFolderText>
           <S.TreeDirectoryTree
             height={height - 2 * DEFAULT_PANE_TITLE_HEIGHT - 20}
-            onSelect={onSelect}
+            onSelect={onFileSelect}
             treeData={[tree]}
             ref={treeRef}
             expandedKeys={expandedFolders}
