+++ src/components/organisms/SearchPane/SearchPane.tsx
@@ -2,46 +2,32 @@ import {ipcRenderer} from 'electron';
 
 import React, {Key, useCallback, useEffect, useMemo, useRef, useState} from 'react';
 import {useSelector} from 'react-redux';
-import {useUpdateEffect} from 'react-use';
 
-import {Input, Modal} from 'antd';
-
-import {ExclamationCircleOutlined} from '@ant-design/icons';
+import {Input} from 'antd';
 
 import log from 'loglevel';
-import micromatch from 'micromatch';
 import path from 'path';
 
 import {DEFAULT_PANE_TITLE_HEIGHT, ROOT_FILE_ENTRY} from '@constants/constants';
 
-import {AlertEnum} from '@models/alert';
 import {FileEntry} from '@models/fileentry';
 
 import {useAppDispatch, useAppSelector} from '@redux/hooks';
-import {setAlert} from '@redux/reducers/alert';
-import {updateProjectConfig} from '@redux/reducers/appConfig';
-import {selectFile, setSelectingFile, updateResourceFilter} from '@redux/reducers/main';
-import {
-  openCreateFolderModal,
-  openNewResourceWizard,
-  openRenameEntityModal,
-  setExpandedFolders,
-} from '@redux/reducers/ui';
-import {fileIncludesSelector, isInPreviewModeSelector, scanExcludesSelector, settingsSelector} from '@redux/selectors';
-import {getResourcesForPath} from '@redux/services/fileEntry';
-import {getHelmValuesFile} from '@redux/services/helm';
-import {isKustomizationResource} from '@redux/services/kustomize';
-import {startPreview, stopPreview} from '@redux/services/preview';
+import {setSelectingFile, updateResourceFilter} from '@redux/reducers/main';
+import {openRenameEntityModal, setExpandedFolders, openCreateFileFolderModal} from '@redux/reducers/ui';
+import {isInPreviewModeSelector, settingsSelector} from '@redux/selectors';
 import {setRootFolder} from '@redux/thunks/setRootFolder';
 
 import {TitleBar} from '@molecules';
 
+import {useCreate, useDelete, useDuplicate, useFileSelect, useHighlightNode, usePreview, useProcessing} from '@hooks/fileTreeHooks';
+
 import electronStore from '@utils/electronStore';
-import {uniqueArr} from '@utils/index';
+import {MatchParamProps, getRegexp} from '@utils/getRegexp';
 
 import {createNode} from '../FileTreePane/CreateNode';
 import TreeItem from '../FileTreePane/TreeItem';
-import {ProcessingEntity, TreeNode} from '../FileTreePane/types';
+import {TreeNode} from '../FileTreePane/types';
 import {createFilteredNode} from './CreateFilteredNode';
 import RecentSearch from './RecentSearch';
 
@@ -51,18 +37,7 @@ type Props = {
   height: number;
 };
 
-type MatchParamProps = {
-  matchCase: boolean;
-  matchWholeWord: boolean;
-  regExp: boolean;
-};
-
 const SearchPane: React.FC<Props> = ({height}) => {
-  const [highlightNode, setHighlightNode] = useState<TreeNode>();
-  const [processingEntity, setProcessingEntity] = useState<ProcessingEntity>({
-    processingEntityID: undefined,
-    processingType: undefined,
-  });
   const [tree, setTree] = useState<TreeNode | null>(null);
   const [searchTree, setSearchTree] = useState<TreeNode | null>(null);
   const [searchQuery, updateSearchQuery] = useState<string>('');
@@ -74,14 +49,11 @@ const SearchPane: React.FC<Props> = ({height}) => {
     regExp: false,
   });
 
-  const leftMenuSelection = useAppSelector(state => state.ui.leftMenu.selection);
   const isInPreviewMode = useSelector(isInPreviewModeSelector);
 
   const dispatch = useAppDispatch();
-  const configState = useAppSelector(state => state.config);
   const fileMap = useAppSelector(state => state.main.fileMap);
   const fileOrFolderContainedInFilter = useAppSelector(state => state.main.resourceFilter.fileOrFolderContainedIn);
-  const helmValuesMap = useAppSelector(state => state.main.helmValuesMap);
   const isFolderLoading = useAppSelector(state => state.ui.isFolderLoading);
   const isSelectingFile = useAppSelector(state => state.main.isSelectingFile);
   const previewLoader = useAppSelector(state => state.main.previewLoader);
@@ -89,14 +61,16 @@ const SearchPane: React.FC<Props> = ({height}) => {
   const resourceMap = useAppSelector(state => state.main.resourceMap);
   const selectedPath = useAppSelector(state => state.main.selectedPath);
   const selectedResourceId = useAppSelector(state => state.main.selectedResourceId);
-
   const {hideExcludedFilesInFileExplorer, hideUnsupportedFilesInFileExplorer} = useAppSelector(settingsSelector);
-  const fileIncludes = useAppSelector(fileIncludesSelector);
-  const scanExcludes = useAppSelector(scanExcludesSelector);
+  const onFileSelect = useFileSelect();
+  const onPreview = usePreview();
+  const {onDelete, processingEntity, setProcessingEntity} = useDelete();
+  const {onDuplicate} = useDuplicate();
+  const {onCreateResource} = useCreate();
 
   const treeRef = useRef<any>();
-
   const expandedFolders = ['filter'];
+  const highlightFilePath = useHighlightNode(tree, treeRef, expandedFolders);
 
   const rootFolderName = useMemo(() => {
     return fileMap[ROOT_FILE_ENTRY] ? path.basename(fileMap[ROOT_FILE_ENTRY].filePath) : ROOT_FILE_ENTRY;
@@ -109,9 +83,15 @@ const SearchPane: React.FC<Props> = ({height}) => {
     [dispatch]
   );
 
+  const onCreateFileFolder = (absolutePath: string, type: 'file' | 'folder') => {
+    dispatch(openCreateFileFolderModal({rootDir: absolutePath, type}));
+  };
+
   const refreshFolder = useCallback(() => {
     setFolder(fileMap[ROOT_FILE_ENTRY].filePath);
   }, [fileMap, setFolder]);
+
+  const {onExcludeFromProcessing, onIncludeToProcessing} = useProcessing(refreshFolder);
 
   useEffect(() => {
     if (isFolderLoading) {
@@ -149,32 +129,6 @@ const SearchPane: React.FC<Props> = ({height}) => {
    * when a resource is selected
    */
 
-  function highlightFilePath(filePath: string) {
-    const paths = filePath.split(path.sep);
-    const keys: Array<React.Key> = [ROOT_FILE_ENTRY];
-
-    for (let c = 1; c < paths.length; c += 1) {
-      keys.push(paths.slice(0, c + 1).join(path.sep));
-    }
-
-    let node: TreeNode | undefined = tree || undefined;
-    for (let c = 1; c < keys.length && node; c += 1) {
-      node = node.children.find((i: any) => i.key === keys[c]);
-    }
-
-    if (node) {
-      node.highlight = true;
-      treeRef?.current?.scrollTo({key: node.key});
-
-      if (highlightNode) {
-        highlightNode.highlight = false;
-      }
-    }
-
-    setHighlightNode(node);
-    dispatch(setExpandedFolders(uniqueArr([...expandedFolders, ...Array.from(keys)])));
-  }
-
   useEffect(() => {
     if (selectedResourceId && tree) {
       const resource = resourceMap[selectedResourceId];
@@ -187,92 +141,8 @@ const SearchPane: React.FC<Props> = ({height}) => {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [selectedResourceId, tree]);
 
-  const onDelete = (args: {isDirectory: boolean; name: string; err: NodeJS.ErrnoException | null}): void => {
-    const {isDirectory, name, err} = args;
-
-    if (err) {
-      dispatch(
-        setAlert({
-          title: 'Deleting failed',
-          message: `Something went wrong during deleting a ${isDirectory ? 'directory' : 'file'}`,
-          type: AlertEnum.Error,
-        })
-      );
-    } else {
-      dispatch(
-        setAlert({
-          title: `Successfully deleted a ${isDirectory ? 'directory' : 'file'}`,
-          message: `You have successfully deleted ${name} ${isDirectory ? 'directory' : 'file'}`,
-          type: AlertEnum.Success,
-        })
-      );
-    }
-
-    /**
-     * Deleting is performed immediately.
-     * The Ant Tree component is not updated immediately.
-     * I show the loader long enough to let the Ant Tree component update.
-     */
-    setTimeout(() => {
-      setProcessingEntity({processingEntityID: undefined, processingType: undefined});
-    }, 2000);
-  };
-
   const onRename = (absolutePathToEntity: string, osPlatform: string) => {
     dispatch(openRenameEntityModal({absolutePathToEntity, osPlatform}));
-  };
-
-  const onSelect = (selectedKeysValue: React.Key[], info: any) => {
-    if (!fileIncludes.some(fileInclude => micromatch.isMatch(path.basename(info.node.key), fileInclude))) {
-      return;
-    }
-    if (scanExcludes.some(scanExclude => micromatch.isMatch(path.basename(info.node.key), scanExclude))) {
-      return;
-    }
-    if (info.node.key) {
-      if (isInPreviewMode) {
-        stopPreview(dispatch);
-      }
-      dispatch(setSelectingFile(true));
-      dispatch(selectFile({filePath: info.node.key}));
-    }
-  };
-
-  const openConfirmModal = () => {
-    Modal.confirm({
-      title: 'You should reload the file explorer to have your changes applied. Do you want to do it now?',
-      icon: <ExclamationCircleOutlined />,
-      cancelText: 'Not now',
-      onOk: () => {
-        refreshFolder();
-      },
-    });
-  };
-
-  const onExcludeFromProcessing = (relativePath: string) => {
-    dispatch(
-      updateProjectConfig({
-        config: {
-          ...configState.projectConfig,
-          scanExcludes: [...scanExcludes, relativePath],
-        },
-        fromConfigFile: false,
-      })
-    );
-    openConfirmModal();
-  };
-
-  const onIncludeToProcessing = (relativePath: string) => {
-    dispatch(
-      updateProjectConfig({
-        config: {
-          ...configState.projectConfig,
-          scanExcludes: scanExcludes.filter(scanExclude => scanExclude !== relativePath),
-        },
-        fromConfigFile: false,
-      })
-    );
-    openConfirmModal();
   };
 
   useEffect(() => {
@@ -302,61 +172,6 @@ const SearchPane: React.FC<Props> = ({height}) => {
       ipcRenderer.removeListener('set-root-folder', onSelectRootFolderFromMainThread);
     };
   }, [onSelectRootFolderFromMainThread]);
-
-  const onCreateFolder = (absolutePath: string) => {
-    dispatch(openCreateFolderModal(absolutePath));
-  };
-
-  const onPreview = useCallback(
-    (relativePath: string) => {
-      const resources = getResourcesForPath(relativePath, resourceMap);
-      if (resources && resources.length === 1 && isKustomizationResource(resources[0])) {
-        startPreview(resources[0].id, 'kustomization', dispatch);
-      } else {
-        const fileEntry = fileMap[relativePath];
-        if (fileEntry) {
-          const valuesFile = getHelmValuesFile(fileEntry, helmValuesMap);
-          if (valuesFile) {
-            startPreview(valuesFile.id, 'helm', dispatch);
-          }
-        }
-      }
-    },
-    [dispatch, fileMap, helmValuesMap, resourceMap]
-  );
-
-  const onCreateResource = ({targetFolder, targetFile}: {targetFolder?: string; targetFile?: string}) => {
-    if (targetFolder) {
-      dispatch(openNewResourceWizard({defaultInput: {targetFolder}}));
-    }
-    if (targetFile) {
-      dispatch(openNewResourceWizard({defaultInput: {targetFile}}));
-    }
-  };
-
-  useEffect(() => {
-    // removes any highlight when a file is selected
-    if (selectedPath && highlightNode) {
-      highlightNode.highlight = false;
-    }
-
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [highlightNode]);
-
-  useUpdateEffect(() => {
-    if (leftMenuSelection !== 'file-explorer') {
-      return;
-    }
-
-    if (selectedPath) {
-      treeRef?.current?.scrollTo({key: selectedPath});
-      return;
-    }
-
-    if (highlightNode) {
-      treeRef?.current?.scrollTo({key: highlightNode.key});
-    }
-  }, [tree]);
 
   const onFilterByFileOrFolder = (relativePath: string | undefined) => {
     dispatch(updateResourceFilter({...resourceFilter, fileOrFolderContainedIn: relativePath}));
@@ -406,22 +221,8 @@ const SearchPane: React.FC<Props> = ({height}) => {
       setSearchTree(null);
       return;
     }
-    let matchParams = 'gi'; // global, case insensitive by default
-    if (queryMatchParam.matchCase) {
-      matchParams = 'g';
-    }
-    if (!queryMatchParam.regExp) {
-      query = query.replace(/[|\\{}()[\]^$+*?.]/g, '\\$&');
-    }
 
-    let queryRegExp = new RegExp(query, matchParams);
-
-    if (queryMatchParam.matchWholeWord) {
-      queryRegExp = new RegExp(`\\b${query}\\b`, matchParams);
-    }
-    if (queryMatchParam.regExp) {
-      queryRegExp = new RegExp(query, matchParams);
-    }
+    const queryRegExp = getRegexp(query, queryMatchParam);
 
     const filteredFileMap: FileEntry[] = Object.values(fileMap)
       .map(file => filterFileMap(file, queryRegExp))
@@ -483,7 +284,7 @@ const SearchPane: React.FC<Props> = ({height}) => {
         {searchTree && (
           <S.TreeDirectoryTree
             height={height - 2 * DEFAULT_PANE_TITLE_HEIGHT - 20}
-            onSelect={onSelect}
+            onSelect={onFileSelect}
             treeData={[searchTree]}
             ref={treeRef}
             expandedKeys={expandedFolders}
@@ -498,7 +299,8 @@ const SearchPane: React.FC<Props> = ({height}) => {
                 onRename={onRename}
                 onExcludeFromProcessing={onExcludeFromProcessing}
                 onIncludeToProcessing={onIncludeToProcessing}
-                onCreateFolder={onCreateFolder}
+                onCreateFileFolder={onCreateFileFolder}
+                onDuplicate={onDuplicate}
                 onCreateResource={onCreateResource}
                 onFilterByFileOrFolder={onFilterByFileOrFolder}
                 onPreview={onPreview}
