+++ src/hooks/fileTreeHooks/index.ts
@@ -1,0 +1,9 @@
+import {useCreate} from './useCreate';
+import {useDelete} from './useDelete';
+import {useFileSelect} from './useFileSelect';
+import {useHighlightNode} from './useHighlightNode';
+import {usePreview} from './usePreview';
+import {useProcessing} from './useProcessing';
+import {useDuplicate} from './useDuplicate';
+
+export {useFileSelect, usePreview, useDelete, useProcessing, useHighlightNode, useCreate, useDuplicate};
